Homework 1, made a calculator with reset button and triangle function.

+++ src/main/java/com/example/demo/models/Adder.java
@@ -1,10 +1,10 @@
 package com.example.demo.models;
 
 public class Adder {
-	int secret1;
+	double secret1;
 	double secret2;
 	
-	public Adder(int first, double second) {
+	public Adder(double first, double second) {
 		secret1 = first;
 		secret2 = second;
